fix(astromer): one first-window embedding per curve in test-data
The shared Astromer test-data generator preprocessed all curves in one padded
batch and stored embeddings[i] for curve[i]. But upstream core/data.get_windows
returns floor(length/max_obs)+1 windows, so a curve whose length is an exact
multiple of max_obs (e.g. any curve truncated to exactly 200 obs) yields a
duplicate extra window. Those extra windows made the embeddings outnumber the
curves and shifted the per-curve pairing, so every curve with >=200 observations
got a wrong reference embedding. The error was masked for the MACHO astromer1/2
parquets (their embeddings are mutually ~1.0 cosine-similar, so the loose test
thresholds still passed) but was exposed by astromer1-ztfdr20.

Fix: embed each light curve independently and keep its first window
(reduction="beginning"), guaranteeing exactly one embedding per curve, aligned
by index. Regenerate astromer1, astromer1-ztfdr20, and astromer2 parquets;
verified each against its ONNX model with reduction="beginning" (all rows
cos >= 0.9999, incl. curves with exactly 200 and >200 obs).

Also:
- Fold astromer1-ztfdr20/surveys.py away: the ZTF-records downloader moves to
  download.py and the curve loader to test_data.py (matching the other models),
  with the records-dir constants in config.py. This also fixes the Python-2
  `except A, B:` clause that crashed the ZTF download under Python 3 (written as
  two separate `except` clauses because ruff 0.15.x's formatter rewrites
  `except (A, B):` back into the invalid form).
- Make the TensorFlow 2.14 dependency installable on arm Linux: add
  tensorflow-cpu-aws for linux/aarch64 (alongside tensorflow for x86_64/Windows
  and tensorflow-macos for macOS arm64) in astromer1, astromer2, and
  astromer1-ztfdr20, and relock.
