+++ models/astromer1-ztfdr20/pyproject.toml
@@ -8,8 +8,9 @@ version = "0.1.0"
 description = "Astromer 1 (ZTF DR20 g-band) ONNX export sub-package"
 requires-python = ">=3.11,<3.12"
 dependencies = [
-    "tensorflow==2.14.*; sys_platform != 'darwin' or platform_machine != 'arm64'",
+    "tensorflow==2.14.*; (sys_platform != 'darwin' or platform_machine != 'arm64') and (sys_platform != 'linux' or platform_machine != 'aarch64')",
     "tensorflow-macos==2.14.*; sys_platform == 'darwin' and platform_machine == 'arm64'",
+    "tensorflow-cpu-aws==2.14.*; sys_platform == 'linux' and platform_machine == 'aarch64'",
     "tf2onnx>=1.16",
     "onnx>=1.15,<1.17",
     "numpy<2",
