+++ models/astromer2/pyproject.toml
@@ -9,8 +9,9 @@ description = "Astromer 2 ONNX export sub-package"
 requires-python = ">=3.11,<3.12"
 dependencies = [
     "joblib",
-    "tensorflow>=2.13,<2.15; sys_platform != 'darwin' or platform_machine != 'arm64'",
+    "tensorflow>=2.13,<2.15; (sys_platform != 'darwin' or platform_machine != 'arm64') and (sys_platform != 'linux' or platform_machine != 'aarch64')",
     "tensorflow-macos>=2.13,<2.15; sys_platform == 'darwin' and platform_machine == 'arm64'",
+    "tensorflow-cpu-aws>=2.13,<2.15; sys_platform == 'linux' and platform_machine == 'aarch64'",
     "tf2onnx>=1.16",
     "onnx>=1.15,<1.17",
     "numpy<2",
